+++ Reflection/Changes.md
@@ -16,6 +16,7 @@
 - ColorContext.forTextComponents doesn't accept parameters anymore (except for the implicit localizer)
 ### New Methods
 - Added .hasSameDimensionsAs(TextDrawContext), .mapFont(...) and .mapColor(...) to TextDrawContext
+- Added .stripControlCharacters to LocalStringLike
 ### Fixes
 - Alignment.position(...) now properly takes parameter bounds position into account
 - Fixed a small bug where Window content size wouldn't update correctly